requirements.cv.txt
opencv-python-headless
easyocr
torch
mediapipe==0.10.14
pandas
opencv-python-headless
easyocr
torch
mediapipe==0.10.14
google-cloud-storage
google-cloud-pubsub
pydantic>=2.0
pandas
python-dotenv
fastapi
uvicorn
